Fixed Navigations
In the package.json file: It is recommended that you add ndoemon to the dev-dependencies.

To add a route to the app. You first to create an express route for it, then you import it the the app.js and add the route the express app.

The view engine(EJS) will render the views and send to the clients. you should not worry abut that.